feat: Add DICOM metadata extraction utilities for various modalities and CLI support
