+++ src/imgtools/dicom/__init__.py
@@ -1,3 +1,4 @@
+# ruff: noqa
 from .dicom_find import find_dicoms
 from .dicom_reader import (
     DicomInput,
@@ -7,6 +8,11 @@ from .dicom_reader import (
 )
 from .read_tags import read_tags
 from .utils import lookup_tag, similar_tags, tag_exists
+from .dicom_metadata.dicom_metadata import (
+    MODALITY_TAGS,
+    extract_dicom_tags,
+    modality_metadata_keys,
+)
 
 __all__ = [
     # dicom_find
@@ -22,4 +28,8 @@ __all__ = [
     "load_seg_dcm",
     # read_tags
     "read_tags",
+    # dicom_metadata
+    "MODALITY_TAGS",
+    "extract_dicom_tags",
+    "modality_metadata_keys",
 ]
