+++ src/imgtools/dicom/dicom_metadata/__main__.py
@@ -1,0 +1,45 @@
+import click
+
+from imgtools.dicom.dicom_metadata import (
+    extract_metadata,
+    supported_modalities,
+)
+
+
+@click.command()
+@click.argument("dicom_file", type=click.Path(exists=True))
+@click.option(
+    "-o",
+    "--output",
+    type=click.Path(writable=True),
+    default=None,
+    help="Output file path for the extracted metadata.",
+)
+@click.option(
+    "--list-modalities",
+    "-lm",
+    is_flag=True,
+    help="List all available DICOM modalities.",
+)
+def cli(dicom_file, output, list_modalities):
+    """Extract DICOM metadata from a file."""
+    if list_modalities:
+        modalities = supported_modalities()
+        click.echo("Available DICOM modalities:")
+        for modality in modalities:
+            click.echo(modality)
+        return
+
+    metadata = extract_metadata(dicom_file)
+
+    if output:
+        from pathlib import Path
+
+        with Path(output).open("w") as f:
+            f.write(str(metadata))
+    else:
+        click.echo(metadata)
+
+
+if __name__ == "__main__":
+    cli()
