+++ src/imgtools/dicom/dicom_metadata/modality_utils/rtstruct_utils.py
@@ -1,0 +1,159 @@
+from typing import TYPE_CHECKING
+
+from imgtools.dicom.dicom_reader import (
+    DicomInput,
+    load_rtstruct_dcm,
+)
+from imgtools.exceptions import (
+    RTSTRUCTAttributeError,
+)
+
+if TYPE_CHECKING:
+    from pydicom.dataset import FileDataset
+
+__all__ = [
+    "extract_roi_meta",
+    "extract_roi_names",
+    "rtstruct_reference_uids",
+    "RTSTRUCTRefSeries",
+    "RTSTRUCTRefStudy",
+    "RTSTRUCTRefSOP",
+]
+
+
+def extract_roi_meta(rtstruct: DicomInput) -> list[dict[str, str]]:
+    """Extract ROI metadata from an RTSTRUCT DICOM file.
+
+    Iterate over the `StructureSetROISequence` in the RTSTRUCT file and extract:
+        - "ROINumber": Unique identifier for the ROI.
+        - "ROIName": Name of the ROI.
+        - "ROIGenerationAlgorithm": Algorithm used to generate the ROI.
+
+    Parameters
+    ----------
+    rtstruct : FileDataset | str | Path | bytes
+        Input RTSTRUCT DICOM dataset or file path.
+
+    Returns
+    -------
+    list of dict[str, str]
+        A list of dictionaries, each containing metadata for an ROI.
+
+    Raises
+    ------
+    RTSTRUCTAttributeError
+        If the RTSTRUCT file does not contain the required `StructureSetROISequence`.
+    """
+
+    dcm_rtstruct: FileDataset = load_rtstruct_dcm(rtstruct)
+    try:
+        roi_sequence = dcm_rtstruct.StructureSetROISequence
+    except AttributeError as e:
+        errmsg = "Failed to extract ROISequence from the RTSTRUCT file."
+        raise RTSTRUCTAttributeError(errmsg) from e
+    roi_metas = []
+    for roi in roi_sequence:
+        roi_meta = {}
+        roi_meta["ROINumber"] = getattr(roi, "ROINumber", "")
+        roi_meta["ROIName"] = getattr(roi, "ROIName", "")
+        roi_meta["ROIGenerationAlgorithm"] = getattr(
+            roi, "ROIGenerationAlgorithm", ""
+        )
+        roi_metas.append(roi_meta)
+    return roi_metas
+
+
+def extract_roi_names(rtstruct: DicomInput) -> list[str]:
+    """Extract a list of ROI names from an RTSTRUCT DICOM file.
+
+    Parameters
+    ----------
+    rtstruct : FileDataset | str | Path | bytes
+        Input RTSTRUCT DICOM dataset or file path.
+
+    Returns
+    -------
+    list of str
+        A list of ROI names extracted from the RTSTRUCT file.
+
+    Raises
+    ------
+    RTSTRUCTAttributeError
+        If the RTSTRUCT file does not contain the required `StructureSetROISequence`.
+    """
+    dcm_rtstruct: FileDataset = load_rtstruct_dcm(rtstruct)
+    try:
+        roi_sequence = dcm_rtstruct.StructureSetROISequence
+    except AttributeError as e:
+        errmsg = "Failed to extract ROISequence from the RTSTRUCT file."
+        raise RTSTRUCTAttributeError(errmsg) from e
+    roi_names = [roi.ROIName for roi in roi_sequence]
+    return roi_names
+
+
+class RTSTRUCTRefSeries(str):
+    pass
+
+
+class RTSTRUCTRefStudy(str):
+    pass
+
+
+class RTSTRUCTRefSOP(str):
+    pass
+
+
+def rtstruct_reference_uids(
+    rtstruct: DicomInput,
+) -> tuple[RTSTRUCTRefSeries, RTSTRUCTRefStudy] | RTSTRUCTRefSOP:
+    """Retrieve the referenced SeriesInstanceUID and StudyInstanceUID from an RTSTRUCT.
+
+    Parameters
+    ----------
+    rtstruct : FileDataset | str | Path | bytes
+        Input RTSTRUCT DICOM dataset or file path.
+
+    Returns
+    -------
+    tuple[RTSTRUCTRefSeries, RTSTRUCTRefStudy]
+        - Referenced `SeriesInstanceUID` (RTSTRUCTRefSeries)
+        - Referenced `StudyInstanceUID` (RTSTRUCTRefStudy)
+
+    Raises
+    ------
+    RTSTRUCTAttributeError
+        If the RTSTRUCT file does not contain the required reference fields.
+    """
+    dcm_rtstruct: FileDataset = load_rtstruct_dcm(rtstruct)
+    try:
+        referenced_series_instance_uid = RTSTRUCTRefSeries(
+            dcm_rtstruct.ReferencedFrameOfReferenceSequence[0]
+            .RTReferencedStudySequence[0]
+            .RTReferencedSeriesSequence[0]
+            .SeriesInstanceUID
+        )
+        referenced_study_instance_uid = RTSTRUCTRefStudy(
+            dcm_rtstruct.ReferencedFrameOfReferenceSequence[0]
+            .RTReferencedStudySequence[0]
+            .ReferencedSOPInstanceUID
+        )
+        return referenced_series_instance_uid, referenced_study_instance_uid
+    except (AttributeError, IndexError) as e:
+        errmsg = "Failed to extract Referenced SeriesInstanceUID or Referenced StudyInstanceUID from the RTSTRUCT file."
+        errmsg += " Attempting to extract ReferencedSOPInstanceUID."
+
+        try:
+            ref_sequence = dcm_rtstruct.ReferencedFrameOfReferenceSequence[0]
+
+            # ref_frame_ref = ref_sequence.FrameOfReferenceUID
+            ref_sop = (
+                ref_sequence.RTReferencedStudySequence[0]
+                .RTReferencedSeriesSequence[0]
+                .ContourImageSequence[0]
+                .ReferencedSOPInstanceUID
+            )
+
+            return RTSTRUCTRefSOP(ref_sop)
+        except (AttributeError, IndexError) as e2:
+            errmsg += f"First error message: {e}. Second error message: {e2}"
+            raise RTSTRUCTAttributeError(errmsg) from e2
